+++ e-commerce/ViewModels/CartVM.cs
@@ -5,7 +5,7 @@ namespace e_commerce.ViewModels
     public class CartVM
     {
         public Cart Cart { get; set; }
-        public decimal CartTotal { get; set; }
+        public double CartTotal { get; set; }
 
     }
 }
